refactor(webapp): one blob download helper instead of three copies
The export made a third copy of the create-object-url, build-an-anchor, click,
revoke dance, and the two existing ones had already drifted - only the flow
download appended the link to the document, which Firefox needs. All three now
call `downloadBlob`.
