+++ packages/webapp/src/utils/download.ts
@@ -1,0 +1,12 @@
+/** Hands the browser a blob to save under `filename`. */
+export function downloadBlob(blob: Blob, filename: string): void {
+    const url = URL.createObjectURL(blob);
+    const link = document.createElement('a');
+    link.href = url;
+    link.download = filename;
+    // Appended before clicking: Firefox ignores a click on a link that is not in the document.
+    document.body.appendChild(link);
+    link.click();
+    document.body.removeChild(link);
+    URL.revokeObjectURL(url);
+}
